ProcessStoreをファイルに切り出し
